requirements.txt
python-dotenv==1.0.0
tavily-python==0.5.0
Pillow==10.1.0
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
